refactor(openapi-mock): restructure src into lib/utils and move tests to __tests__
Splits the previously flat src/ into lib/ (public building blocks) and utils/ (shared internal helpers), deduplicates resolvePointer, moves tests from tests/ to the conventional src/__tests__/, and documents the new seeded faker field on FieldFakerContext.
