+++ packages/utils/openapi-mock/src/__tests__/fixtures/fields-json.faker.json
