+++ packages/utils/openapi-mock/src/__tests__/fixtures/fields-ts.faker.ts
@@ -1,0 +1,6 @@
+import type { FieldFakerContext, FieldFakerMap } from '../../types.js';
+
+export default {
+  'User.email': 'internet.email',
+  'User.id': ({ modelName, path }: FieldFakerContext) => `${modelName}:${path.join('.')}`,
+} satisfies FieldFakerMap;
