+++ packages/utils/openapi-mock/src/__tests__/fixtures/fields-yaml.faker.yaml
