+++ packages/utils/openapi-mock/src/lib/generate-mock-from-schema.ts
@@ -1,8 +1,6 @@
 import { Faker, en } from '@faker-js/faker';
 import { generate } from 'json-schema-faker';
 
-import type { FieldFakerFn } from './types';
-
 /** Options for {@link generateMockFromSchema}. */
 export interface GenerateMockFromSchemaOptions {
   /**
@@ -15,9 +13,10 @@ export interface GenerateMockFromSchemaOptions {
   /**
    * Function-valued {@link FieldFakerMap} entries, keyed by the synthetic
    * `"__custom.<key>"` faker path {@link applyFieldFakers} annotated the
-   * schema with, so those paths resolve during generation.
+   * schema with, awaiting only the seeded `faker` instance this function
+   * creates before they resolve during generation.
    */
-  customFakers?: Record<string, FieldFakerFn>;
+  customFakers?: Record<string, (faker: Faker) => unknown>;
 }
 
 /**
@@ -40,10 +39,14 @@ export async function generateMockFromSchema(
   const localFaker = new Faker({ locale: [en] });
   // Only a caller-supplied seed makes output deterministic; otherwise this instance stays random.
   if (options.seed !== undefined) localFaker.seed(options.seed);
+  // Each pending custom faker only needed this seeded instance to become callable.
+  const customFakers = Object.fromEntries(
+    Object.entries(options.customFakers ?? {}).map(([key, fn]) => [key, () => fn(localFaker)]),
+  );
   return generate(schema as Parameters<typeof generate>[0], {
     // `__custom` is a synthetic namespace `applyFieldFakers` points function-valued
     // field overrides at, resolved by the same dotted-path lookup as a real faker path
-    extensions: { faker: { ...localFaker, __custom: options.customFakers ?? {} } },
+    extensions: { faker: { ...localFaker, __custom: customFakers } },
     seed: options.seed,
     // fake every optional property too, so a consumer sees the whole schema shape rather
     // than a coin-flip subset of it on every run
