--- packages/utils/openapi-mock/tests/fixtures/fields-ts.faker.ts
@@ -1,6 +1,0 @@
-import type { FieldFakerMap } from '../../src/types';
-
-export default {
-  'User.email': 'internet.email',
-  'User.id': ({ modelName, path }) => `${modelName}:${path.join('.')}`,
-} satisfies FieldFakerMap;
