+++ packages/utils/openapi-mock/tsconfig.json
@@ -1,6 +1,8 @@
 {
   "extends": "../../../tsconfig.base.json",
   "compilerOptions": {
+    "module": "NodeNext",
+    "moduleResolution": "nodenext",
     "outDir": "dist/esm",
     "rootDir": "src",
     "declarationDir": "./dist/types"
@@ -10,6 +12,6 @@
       "path": "../imports"
     }
   ],
-  "include": ["src/**/*", "tests/**/*"],
-  "exclude": ["node_modules", "tests"]
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "__tests__"]
 }
